requirements.txt
customtkinter
python-docx
pyinstaller